+++ src/main/kotlin/com/rickbusarow/ktrules/KtRulesRuleSetProvider.kt
@@ -23,6 +23,7 @@ import com.pinterest.ktlint.core.RuleSetProviderV2
 import com.rickbusarow.ktrules.rules.KDocIndentAfterLeadingAsteriskRule
 import com.rickbusarow.ktrules.rules.KDocLeadingAsteriskRule
 import com.rickbusarow.ktrules.rules.KDocTagOrderRule
+import com.rickbusarow.ktrules.rules.KDocTagParamOrPropertyRule
 import com.rickbusarow.ktrules.rules.KDocWrappingRule
 import com.rickbusarow.ktrules.rules.NoDuplicateCopyrightHeaderRule
 import com.rickbusarow.ktrules.rules.NoLeadingBlankLinesRule
@@ -48,6 +49,7 @@ class KtRulesRuleSetProvider : RuleSetProviderV2(
       RuleProvider { KDocIndentAfterLeadingAsteriskRule() },
       RuleProvider { KDocLeadingAsteriskRule() },
       RuleProvider { KDocTagOrderRule() },
+      RuleProvider { KDocTagParamOrPropertyRule() },
       RuleProvider { KDocWrappingRule() },
       RuleProvider { NoDuplicateCopyrightHeaderRule() },
       RuleProvider { NoLeadingBlankLinesRule() },